Chore: 폴더, 파일 분리
