+++ src/js/constants/constants.js
@@ -1,0 +1,6 @@
+export const PLUS = "+";
+export const MINUS = "-";
+export const MULTIPLICATION = "X";
+export const DIVISION = "/";
+
+export const ALERT_MESSAGE = "숫자는 한번에 최대 3자리 수까지 입력 가능합니다!";
